myapp/requirements.txt
anyascii==0.3.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.3
django-recaptcha==3.0.0
django-restframework==0.0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.3
html5lib==1.1
idna==3.3
ipython==8.2.0
jedi==0.18.1
l18n==2021.3
matplotlib-inline==0.1.3
oauthlib==3.2.0
openpyxl==3.0.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.0
telepath==0.2
traitlets==5.1.1
urllib3==1.26.9
wagtail==2.16.1
wagtail-django-recaptcha==1.0
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
